Use the Collection and Document to build gallery
This now creates an `index.html` for each gallery folder, even if there
is no input file. A document for the gallery also has the list of images
that are attached to that gallery as a property that can be accessed
from Liquid.

This makes the code a lot nicer and has the same functionality as before,
only actually working: with a slight change to the gallery layout, it now
displays images and fits into the layout of the theme!